pyproject.toml
[tool.poetry]
name = "mosar-graphrag"
version = "0.1.0"
description = "MOSAR Requirements Engineering GraphRAG System"
authors = ["MOSAR Team <mosar@spaceapps.co.kr>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Core
neo4j = "^5.14.0"
langgraph = "^0.2.16"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-core = "^0.3.0"

# NLP & Embeddings
openai = "^1.50.0"
spacy = "^3.7.0"
spacy-transformers = "^1.3.0"
sentence-transformers = "^2.3.0"

# Entity Resolution
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.25.0"

# Utilities
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.1.0"
jupyter = "^1.0.0"
ipykernel = "^6.28.0"
langsmith = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
